backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
psutil>=5.9.0
pillow>=10.0.0
pydantic>=2.0.0
docker>=7.1.0
aiosqlite>=0.19.0
databases>=0.8.0
bcrypt==3.2.2
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-dotenv==1.0.1